refactor: change test and build setup
