+++ .github/workflows/ci.yaml
@@ -28,5 +28,8 @@ jobs:
       - name: Run build
         run: npm run build
 
+      - name: Check exports
+        run: npm run check-exports
+
       - name: Run tests
         run: npm run test
