+++ demo/validation/has-shape-of.ts
