+++ package.json
@@ -1,25 +1,34 @@
 {
-  "name": "express-decorators",
+  "name": "clean-server",
   "version": "0.1.0",
-  "description": "",
-  "main": "src/index.ts",
-  "scripts": {
-    "serve": "ts-node src/index.ts",
-    "test": "jest",
-    "build": "tsc"
+  "description": "A simpler, cleaner way to define your APIs in typescript.",
+  "type": "module",
+  "main": "dist/index.js",
+  "homepage": "https://github.com/sharkdarras/clean-server",
+  "bugs": {
+    "url": "https://github.com/sharkdarras/clean-server/issues"
   },
-  "author": "",
-  "license": "ISC",
+  "scripts": {
+    "test": "vitest run",
+    "build": "tsc",
+    "check-exports": "attw --pack . --ignore-rules=cjs-resolves-to-esm"
+  },
+  "author": "Charles Darras <info@charlesdarras.com> (https://charlesdarras.com)",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sharkdarras/clean-server.git"
+  },
+  "license": "MIT",
   "dependencies": {
-    "express": "^4.21.2",
     "runtypes": "^7.0.4"
   },
   "devDependencies": {
-    "@types/express": "^5.0.0",
-    "@types/jest": "^29.5.14",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.2.6",
+    "@arethetypeswrong/cli": "^0.17.4",
     "ts-node": "^10.9.2",
-    "typescript": "^5.8.2"
-  }
+    "typescript": "^5.8.2",
+    "vitest": "^3.0.7"
+  },
+  "files": [
+    "dist"
+  ]
 }
