+++ src/errors/error-mapper.ts
@@ -1,4 +1,4 @@
-import { HttpResponse } from "../http/http-response";
+import { type HttpResponse } from "../http/http-response.js";
 
 export interface ErrorMapper<TErrorResponseBody> {
   mapToHttpResponse(error: unknown): HttpResponse<TErrorResponseBody>;
