+++ src/index.ts
@@ -1,14 +1,13 @@
-import { CreateExpenseRequestHandler } from "./demo/create-expense-request-handler";
-import { GetExpenseRequestHandler } from "./demo/get-expense-request-handler";
-import { CleanServer } from "./lib/clean-server";
-
-const PORT = parseInt(process.env["PORT"] || "3000");
-
-const server = new CleanServer();
-
-server.addRequestHandler(new GetExpenseRequestHandler());
-server.addRequestHandler(new CreateExpenseRequestHandler());
-
-server.start(PORT, () => {
-  console.log(`Server is running on http://localhost:${PORT}`);
-});
+export * from "./body/body-parser.js";
+export * from "./body/json-only-body-parser.js";
+export * from "./clean-server.js";
+export * from "./errors/error-mapper.js";
+export * from "./http/http-headers.js";
+export * from "./http/http-method.js";
+export * from "./http/http-request.js";
+export * from "./http/http-response.js";
+export * from "./path/path.js";
+export * from "./request-handler.js";
+export * from "./sanitizing/passthrough-sanitizer.js";
+export * from "./sanitizing/runtype-sanitizer.js";
+export * from "./sanitizing/sanitizer.js";
