+++ src/path/path-segment.ts
