+++ src/sanitizing/passthrough-sanitizer.ts
@@ -1,4 +1,4 @@
-import { Sanitizer } from "./sanitizer";
+import { type Sanitizer } from "./sanitizer.js";
 
 export class PassthroughSanitizer implements Sanitizer<any> {
   sanitize(value: unknown): any {
