+++ src/sanitizing/runtype-sanitizer.ts
@@ -1,4 +1,4 @@
-import { Sanitizer } from "./sanitizer";
+import { type Sanitizer } from "./sanitizer.js";
 
 export class RuntypeSanitizer<TParsed> implements Sanitizer<TParsed> {
   constructor(private runtype: RuntypeChecker<TParsed>) {}
